+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Terminate the process when one of the services crashes. ([#4249](https://github.com/getsentry/relay/pull/4249))
 - Don't propagate trace sampling decisions from SDKs ([#4265](https://github.com/getsentry/relay/pull/4265))
 
+**Features**:
+
+- Implement zstd http encoding for Relay to Relay communication. ([#4266](https://github.com/getsentry/relay/pull/4266))
+
 ## 24.11.0
 
 **Breaking Changes**: